refactor(utils): move `StaticVector` definition to its own file
Bench: 5576530
